// src/app/api/model/product.ts
/**
 * E-Commerce API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface Product {
  id: string;
  name: string;
  description?: string;
  price: number;
  currency: string;
  images?: Array<string>;
  category?: string;
  stock?: number;
  rating?: number;
}